+++ bindings/nostr-sdk-js/src/database.rs
@@ -54,6 +54,7 @@ impl JsNostrDatabase {
     pub async fn save_event(&self, event: &JsEvent) -> Result<bool> {
         self.inner.save_event(event).await.map_err(into_err)
     }
+
     /// Get list of relays that have seen the [`EventId`]
     #[wasm_bindgen(js_name = eventSeenOnRelays)]
     pub async fn event_seen_on_relays(